// src/enums/modelNames.enum.ts
export enum ModelNamesEnum {
  CART = 'CART',
  CUSTOMER = 'CUSTOMER',
  DESCRIPTION = 'DESCRIPTION',
  DOMAIN = 'DOMAIN',
  ORDER = 'ORDER',
  CATEGORY = 'CATEGORY',
  COMPANY = 'COMPANY',
  IMAGE = 'IMAGE',
  VARIATION = 'VARIATION',
  PRODUCT = 'PRODUCT',
  DETAILS = 'DETAILS',
  PROJECT = 'PROJECT',
  SEO = 'SEO',
  USER = 'USER',
}